requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
jinja2
python-jose
email-validator
starlette
PyJWT
itsdangerous
python-multipart
spacy
dateparser
en_core_web_sm
twilio